Move 2 tests
