+++ ecmascript/transforms/compat/tests/es2015_function_name.rs
@@ -258,90 +258,6 @@ const z = {
 "#
 );
 
-// function_name_modules
-test!(
-    syntax(),
-    |_| chain!(
-        resolver(),
-        decorators(decorators::Config {
-            legacy: true,
-            ..Default::default()
-        }),
-        classes(),
-        function_name(),
-        common_js(Mark::fresh(Mark::root()), Default::default()),
-    ),
-    function_name_modules,
-    r#"
-import events from "events";
-
-class Template {
-  events() {
-    return events;
-  }
-}
-
-console.log(new Template().events());
-
-"#,
-    r#"
-"use strict";
-
-var _events = _interopRequireDefault(require("events"));
-
-let Template =
-/*#__PURE__*/
-function () {
-  'use strict';
-  function Template() {
-    _classCallCheck(this, Template);
-  }
-
-  _createClass(Template, [{
-    key: "events",
-    value: function events() {
-      return _events.default;
-    }
-  }]);
-  return Template;
-}();
-
-console.log(new Template().events());
-
-"#
-);
-
-// function_name_eval
-test!(
-    syntax(),
-    |_| chain!(
-        resolver(),
-        function_name(),
-        classes(),
-        decorators(decorators::Config {
-            legacy: true,
-            ..Default::default()
-        })
-    ),
-    function_name_eval,
-    r#"
-var a = {
-  eval: function () {
-    return eval;
-  }
-};
-
-"#,
-    r#"
-var a = {
-  eval: function _eval() {
-    return eval;
-  }
-};
-
-"#
-);
-
 // function_name_modules_3
 test!(
     syntax(),
